+++ src/main/java/kr/co/monitoringserver/service/service/user/UserService.java
@@ -9,8 +9,8 @@ import kr.co.monitoringserver.persistence.entity.User;
 import kr.co.monitoringserver.persistence.repository.AttendanceRepository;
 import kr.co.monitoringserver.persistence.repository.UserAttendanceRepository;
 import kr.co.monitoringserver.service.dtos.request.UserAttendanceReqDTO;
-import kr.co.monitoringserver.service.dtos.request.UserRequestDto;
 import kr.co.monitoringserver.persistence.repository.UserRepository;
+import kr.co.monitoringserver.service.dtos.request.UserReqDTO;
 import kr.co.monitoringserver.service.dtos.response.UserAttendanceResDTO;
 import kr.co.monitoringserver.service.enums.AttendanceType;
 import kr.co.monitoringserver.service.enums.RoleType;
@@ -46,12 +46,12 @@ public class UserService {
     private final UserAttendanceMapper userAttendanceMapper;
 
     @Transactional
-    public void join(UserRequestDto userDto) {
+    public void join(UserReqDTO userDto) {
         User user = User.builder()
                 .identity(userDto.getIdentity())
                 .password(encoder.encode(userDto.getPassword()))
                 .name(userDto.getName())
-                .telephone(userDto.getPhone())
+                .telephone(userDto.getTelephone())
                 .department(userDto.getDepartment())
                 .roleType(RoleType.USER1)
                 .build();
@@ -71,7 +71,7 @@ public class UserService {
     }
 
     @Transactional
-    public void update(UserRequestDto userDto) {
+    public void update(UserReqDTO userDto) {
         User persistance = userRepository.findByIdentity(userDto.getIdentity())
                 .orElseThrow(()->{
                     return new IllegalArgumentException("회원 찾기 실패");
@@ -82,7 +82,7 @@ public class UserService {
         persistance.setPassword(encPassword);
         persistance.setName(userDto.getName());
         persistance.setDepartment(userDto.getDepartment());
-        persistance.setTelephone(userDto.getPhone());
+        persistance.setTelephone(userDto.getTelephone());
     }
